refactor(builtins): split functions.rs into cohesive submodules (§7-8)
Split the 2103-line builtins/functions.rs (over the 500-line limit) into a
facade plus eleven themed submodules under src/builtins/functions/. All but
one are under 500 lines; dispatch_1arg.rs (775) holds the single ~754-line
native_function_1arg function, which cannot be split at a function boundary.
functions.rs is free functions addressed as functions::<name> (externally via
builtins::<name> re-exports); the facade re-exports each submodule item
preserving its original visibility, so no external caller changes.
src/builtins/mod.rs is untouched.

- functions.rs (83)          - facade: mod decls, re-exports, native_function entry
- dispatch_0arg.rs (21)      - native_function_0arg
- dispatch_1arg.rs (775)     - native_function_1arg (single indivisible fn)
- dispatch_2arg.rs (426)     - native_function_2arg
- dispatch_3arg.rs (50)      - native_function_3arg
- dispatch_variadic.rs (381) - native_function_variadic
- flat.rs (151)              - is_infinite_range, deitemize_flat_operand, flat_val, join_flat
- junction.rs (37)           - build_junction
- math.rs (58)               - gcd_u64, factorial_bigint, generic_range_as_bigint, is_extrema_named_pair
- sprintf_fmt.rs (50)        - native_sprintf
- time.rs (113)              - builtin_times, builtin_localtime_gmtime
- uniparse.rs (39)           - uniparse_impl

Pure relocation. Moved fns are pub(crate) in submodules and re-exported at
their original visibility. super:: body refs rewritten to crate::builtins::
paths; cross-sibling refs use super::sibling_mod:: paths. The fn multiset is
byte-identical. No behavior change.

External functions:: callers preserved without edits:
- src/builtins/mod.rs: build_junction, join_flat, native_function, deitemize_flat_operand, flat_val
- src/builtins/methods_0arg/collection.rs: factorial_bigint
- src/builtins/methods_0arg/dispatch_core_unicode.rs: uniparse_impl
- src/runtime/builtins_collection.rs: comment reference only
